aaa: The SSO token expiration date should be returned as Long
The SSO token expiration needs to returned as
a long value and should not be cast to int.

The token expiration is the Authrization
record expiration date returned from the
ldap server and should not be confused with
engine session expiration date.

Bug-Url: https://bugzilla.redhat.com/1391643

+++ backend/manager/modules/enginesso/src/main/java/org/ovirt/engine/core/sso/utils/SsoSession.java
@@ -22,7 +22,7 @@ public class SsoSession implements Serializable {
     private String appUrl;
     private String authorizationCode;
     private String accessToken;
-    private int validTo;
+    private long validTo;
     private String userId;
     private String scope;
     private List<String> scopeAsList;
@@ -107,11 +107,11 @@ public class SsoSession implements Serializable {
         this.appUrl = appUrl;
     }
 
-    public int getValidTo() {
+    public long getValidTo() {
         return validTo;
     }
 
-    public void setValidTo(int validTo) {
+    public void setValidTo(long validTo) {
         this.validTo = validTo;
     }
